GlobalDB consistency test should have the version set
