+++ rotkehlchen/tests/fixtures/__init__.py
@@ -1,6 +1,7 @@
 from rotkehlchen.tests.fixtures.accounting import *  # noqa: F403
 from rotkehlchen.tests.fixtures.assets import *  # noqa: F403
 from rotkehlchen.tests.fixtures.blockchain import *  # noqa: F403
+from rotkehlchen.tests.fixtures.dataupdates import *  # noqa: F403
 from rotkehlchen.tests.fixtures.db import *  # noqa: F403
 from rotkehlchen.tests.fixtures.eth2 import *  # noqa: F403
 from rotkehlchen.tests.fixtures.exchanges import *  # noqa: F403
